Adding example usage of contiguousBlocks and adding
to performance/iterator timing test.

In simple test this approach appears to perform extremely well at the
expense of a slightly more ugly loop construct.

There's also a question of when these blocks would be constructed and
where the correct place to put this getContiguousBlocks routine is
stored. I've assumed that this would be done at initialisation time and
hence haven't included it in the timing.

I've also added a field3D::operator[] that takes an int for this to
actually work -- we could of course change the type of the inner loop
to lock this down a bit.